Basic car movement setup
I have created a physics based movement system for the hover cars. Currently 4 raycasts that are shot from different points towards the floor calulate the compression amount for suspension and apply a force to push the car off the floor based on the distance. The closer to the floor the more force etc.