Controller Support & HUD Update
I have changed the way that the controls work for the controller. Now to move forward the player needs to hold the right trigger and to reverse, the player needs to hold the left trigger. Also assigned the right shoulder button to use the current ability, and updated the HUD for each player to display the correct amount of damage that they currently have.